requirements.txt
python-gitlab
semver
requests
